requirements.txt
streamlit
pytesseract
Pillow
googlesearch-python
requests
beautifulsoup4
